requirements.txt
asgiref==3.7.2
Django==5.0.3
django-cors-headers==4.3.1
django-phonenumber-field==8.0.0
djangorestframework==3.15.1
phonenumbers==8.13.39
phonenumberslite==8.13.39
pillow==10.2.0
sqlparse==0.4.4
tzdata==2024.1
